requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
python-decouple>=3.8
django-extensions

# Image handling
Pillow>=10.0.0

# Database (optional: only if using PostgreSQL)
psycopg2-binary>=2.9.9
